refacturing postreco.c
